+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Check that the Lizmap Web Client target version exists in the table before saving the CFG file
 * Improve the CFG generation when the attribute table has a custom configuration
 * Always export the new checkbox for a fixed scale in the overview map
 * Remove the code related to the server side. It's a new plugin called "Lizmap server" now.